fix: kill container after calling eviction api success

+++ pkg/koordlet/qosmanager/plugins/cpuevict/cpu_evict.go
@@ -292,24 +292,24 @@ func (c *cpuEvictor) killAndEvictBEPodsRelease(node *corev1.Node, bePodInfos []*
 		node.Name, cpuNeedMilliRelease)
 
 	cpuMilliReleased := int64(0)
-	var killedPods []*corev1.Pod
+	hasKillPods := false
 	for _, bePod := range bePodInfos {
 		if cpuMilliReleased >= cpuNeedMilliRelease {
 			break
 		}
 
-		podKillMsg := fmt.Sprintf("%s, kill pod: %s", message, util.GetPodKey(bePod.pod))
-		helpers.KillContainers(bePod.pod, podKillMsg)
+		ok := c.evictor.EvictPodIfNotEvicted(bePod.pod, node, resourceexecutor.EvictPodByBECPUSatisfaction, message)
+		if ok {
+			podKillMsg := fmt.Sprintf("%s, kill pod: %s", message, util.GetPodKey(bePod.pod))
+			helpers.KillContainers(bePod.pod, podKillMsg)
 
-		killedPods = append(killedPods, bePod.pod)
-		cpuMilliReleased = cpuMilliReleased + bePod.milliRequest
-
-		klog.V(5).Infof("cpuEvict pick pod %s/%s to evict", util.GetPodKey(bePod.pod))
+			cpuMilliReleased = cpuMilliReleased + bePod.milliRequest
+			klog.V(5).Infof("cpuEvict pick pod %s to evict", util.GetPodKey(bePod.pod))
+			hasKillPods = true
+		}
 	}
 
-	c.evictor.EvictPodsIfNotEvicted(killedPods, node, resourceexecutor.EvictPodByBECPUSatisfaction, message)
-
-	if len(killedPods) > 0 {
+	if hasKillPods {
 		c.lastEvictTime = time.Now()
 	}
 	klog.V(5).Infof("killAndEvictBEPodsRelease finished! cpuNeedMilliRelease(%d) cpuMilliReleased(%d)",
